+++ www/vavl03/sp-vavl03/cart.php
@@ -21,6 +21,7 @@ if (is_array($ids) && count($ids)) {
         }
         array_push($productsToShow, $product);
     }
+    $_SESSION['productsToShow'] = $productsToShow; //needed for updating product pcs in cart, when changing pcs
     $productSums = [];
     foreach ($productsToShow as $product) {
         $productSum = $product['product_price'] * (int)$product['product_pcs'];
@@ -92,51 +93,4 @@ function getProductPrice($productPrice, $productPcs)
 </main>
 
 <?php require './incl/footer.php'; ?>
-
-<script>
-    // prevent pasting any values in pieces input
-    $("[type='number']").keypress(function(evt) {
-        evt.preventDefault();
-    });
-    // disable products with only 1 piece
-    $(".pcs-input").each((i, e) => {
-        const max = $(e).attr('max');
-        if (max == 1) {
-            $(e).prop('disabled', true);
-        }
-    });
-    // Update live prices on page
-    const products = <?php echo json_encode($productsToShow) ?>;
-    $(".pcs-input").bind('keyup mouseup', async function() {
-        //update individual product price
-        const pcs = $(this).val();
-        const name = $(this).siblings("h5").text();
-        const productData = await jQuery.ajax({
-            url: 'https://eso.vse.cz/~vavl03/sp-vavl03/components/getProduct.php',
-            type: 'GET',
-            data: {
-                productName: name,
-                productPcs: pcs,
-                productsInCart: products
-            },
-            dataType: 'json',
-            success: function(status, status_message, data) {
-                console.log(status)
-            },
-            error: function(status, status_message, data) {
-                console.log(status);
-            }
-        });
-        const productPrice = productData.data;
-        const newPrice = productPrice * pcs;
-        $(this).siblings(".product-price").text(newPrice + " $");
-        // update total price
-        let productPrices = [];
-        document.querySelectorAll(".product-price").forEach((list) => {
-            productPrices.push(parseInt(list.innerHTML.split(".")[0]));
-        });
-        const newTotalPrice = productPrices.reduce((a, b) => a + b, 0);
-        $('.total-price-number').text(newTotalPrice);
-
-    })
-</script>
+<script src="js/cartPrices.js"></script>
