+++ www/vavl03/sp-vavl03/css/styles.css
@@ -415,7 +415,10 @@ a {
     background: #333;
     background-image: linear-gradient(to right, #ccc, #333, #ccc);
 }
-
+.my-orders .no-orders{
+    margin-bottom: 530px;
+    text-align: center;
+}
 .navbar-collapse {
     justify-content: flex-end;
 }
