+++ www/vavl03/sp-vavl03/order_methods.php
@@ -42,25 +42,4 @@ if (isset($_SESSION['orderSent'])) {
 </div>
 <?php require __DIR__ . '/incl/footer.php' ?>
 
-<script>
-    // enable/disable options
-    if ($('#delivery-method').val() === 'homeDelivery') {
-        $("option[value='cashOnPersonalCollection']").attr("disabled", "disabled");
-    }
-    if ($('#delivery-method').val() === 'personalCollection') {
-        $("option[value='cashOnHomeDelivery']").attr("disabled", "disabled");
-    }
-
-    $('#delivery-method').on('change', function() {
-        console.log($(this).val());
-        if ($(this).val() === 'personalCollection') {
-            $('#payment-method').val('bankTransfer');
-            $("option[value='cashOnHomeDelivery']")
-                .attr("disabled", "disabled").siblings().removeAttr("disabled");
-        } else if ($(this).val() === 'homeDelivery') {
-            $('#payment-method').val('bankTransfer');
-            $("option[value='cashOnPersonalCollection']")
-                .attr("disabled", "disabled").siblings().removeAttr("disabled");
-        }
-    });
-</script>
+<script src="js/orderMethods.js"></script>
